+++ src/main/java/reminderebot/ReminderebotException.java
@@ -1,8 +1,10 @@
+package reminderebot;
+
 /**
  * Custom Exception for Reminderebot.
  */
 public class ReminderebotException extends Exception{
-    ReminderebotException(String message) {
+    public ReminderebotException(String message) {
         super("Oops! " + message);
     }
 }
