+++ src/main/java/reminderebot/commands/DeadlineCommand.java
@@ -1,3 +1,13 @@
+package reminderebot.commands;
+
+import reminderebot.TaskList;
+import reminderebot.Ui;
+import reminderebot.Storage;
+import reminderebot.ReminderebotException;
+import reminderebot.task.Task;
+import reminderebot.task.ToDo;
+import reminderebot.task.Deadline;
+import reminderebot.task.Event;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
