+++ src/main/java/reminderebot/commands/DeleteCommand.java
@@ -1,3 +1,15 @@
+package reminderebot.commands;
+
+import reminderebot.TaskList;
+import reminderebot.Ui;
+import reminderebot.Storage;
+import reminderebot.ReminderebotException;
+import reminderebot.task.Task;
+import reminderebot.task.ToDo;
+import reminderebot.task.Deadline;
+import reminderebot.task.Event;
+
+
 /**
  * The DeleteCommand class represents a command to delete a task in tasklist.
  */
