+++ src/main/java/reminderebot/commands/UnmarkCommand.java
@@ -1,19 +1,26 @@
+package reminderebot.commands;
+
+import reminderebot.TaskList;
+import reminderebot.Ui;
+import reminderebot.Storage;
+import reminderebot.ReminderebotException;
+
 /**
- * The MarkCommand class represents a command to mark a task in tasklist as done.
+ * The UnmarkCommand class represents a command to mark a task in tasklist as undone.
  */
-public class MarkCommand extends Command {
+public class UnmarkCommand extends Command {
     private final int index;
 
     /**
-     * Create a Mark command.
+     * Create an Unmark command.
      * @param index
      */
-    public MarkCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
     /**
-     * Marks task as done.
+     * Marks task as undone.
      * @param tasklist
      * @param ui
      * @param storage
@@ -23,14 +30,14 @@ public class MarkCommand extends Command {
     public void execute(TaskList tasklist, Ui ui, Storage storage) throws ReminderebotException {
         if (index > tasklist.length() || index < 1) { // index is out of bounds
             throw new ReminderebotException("Item selected to be marked is not in list.\n" +
-                    "Syntax: mark <int>");
+                    "Syntax: unmark <int>");
         }
-        tasklist.markTask(index);
-        ui.markTask(tasklist.getTask(index - 1));
+        tasklist.unmarkTask(index);
+        ui.unmarkTask(tasklist.getTask(index - 1));
     }
 
     /**
-     * Mark command does not exit.
+     * Unmark command does not exit.
      * @return false
      */
     @Override
