+++ src/main/java/reminderebot/task/Event.java
@@ -1,3 +1,5 @@
+package reminderebot.task;
+
 import java.time.LocalDateTime;
 
 /**
