+++ src/main/java/reminderebot/task/ToDo.java
@@ -1,3 +1,5 @@
+package reminderebot.task;
+
 /**
  * ToDo represents a concrete Task with a description.
  */
